+++ app/Http/Controllers/Auth/ResetPasswordController.php
@@ -2,11 +2,13 @@
 
 namespace Thinktomorrow\Chief\App\Http\Controllers\Auth;
 
+use Illuminate\Support\Facades\Password;
 use Thinktomorrow\Chief\App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Foundation\Auth\ResetsPasswords;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Password;
+use Thinktomorrow\Chief\Authorization\ChiefPasswordBroker;
+use Thinktomorrow\Chief\Authorization\ChiefPasswordBrokerResolver;
 
 class ResetPasswordController extends Controller
 {
@@ -23,17 +25,9 @@ class ResetPasswordController extends Controller
 
     use ResetsPasswords;
 
-    /**
-     * Where to redirect users after resetting their password.
-     * @var string
-     */
-    protected $redirectTo;
-
     public function __construct()
     {
         $this->middleware('chief-guest');
-
-        $this->redirectTo = route('chief.back.dashboard');
     }
 
     public function showResetForm(Request $request, $token = null)
@@ -50,6 +44,34 @@ class ResetPasswordController extends Controller
 
     protected function broker()
     {
-        return Password::broker('chief');
+        return (new ChiefPasswordBrokerResolver(app()))->resolve();
+    }
+
+    public function redirectTo()
+    {
+        return route('chief.back.dashboard');
+    }
+
+    // Override the reset method because chief uses different lang keys and
+    // laravel internals expects this to be of a specific value.
+    public function reset(Request $request)
+    {
+        $request->validate($this->rules(), $this->validationErrorMessages());
+
+        // Here we will attempt to reset the user's password. If it is successful we
+        // will update the password on an actual user model and persist it to the
+        // database. Otherwise we will parse the error and return the response.
+        $response = $this->broker()->reset(
+            $this->credentials($request), function ($user, $password) {
+            $this->resetPassword($user, $password);
+        }
+        );
+
+        // If the password was successfully reset, we will redirect the user back to
+        // the application's home authenticated view. If there is an error we can
+        // redirect them back to where they came from with their error message.
+        return $response == ChiefPasswordBroker::PASSWORD_RESET
+            ? $this->sendResetResponse($request, $response)
+            : $this->sendResetFailedResponse($request, $response);
     }
 }